% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{check_heteroatoms}
\alias{check_heteroatoms}
\title{Check Heteroatoms}
\usage{
check_heteroatoms(heteroatoms)
}
\arguments{
\item{heteroatoms}{string vector of heteroatoms for by_all table.}
}
\value{
proper heteroatom list, dropping non-elements and including CH
}
\description{
Make sure all atoms requested to be in summary table are true atoms, and not other words, or non-character. Allows for Deuterium (D) atoms.
}
\details{
List of allowable elements can be seen by calling \code{GCMSResultsCompiler::elements}
}
\keyword{Internal}
