% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportCompiler.R
\name{compile_reports}
\alias{compile_reports}
\title{Compile Results & Generate a Report}
\usage{
compile_reports(source_file_dir, results_files_dir = NULL,
  report_template = NULL, heteroatoms = GCMSResultsCompiler::elements,
  write_pdf = TRUE, write_csv = TRUE, verbose = 1, ...)
}
\arguments{
\item{source_file_dir}{The folder containing the source MassHunter .xlsx files}

\item{results_files_dir}{The folder in which to put results. Defaults to \code{source_file_dir}/results/}

\item{report_template}{The report template. A default template is included in the package under /rmd/genericreport.rmd}

\item{heteroatoms}{The list of atoms and heteroatoms to check for in the 'all atoms' table. Must include C, H. Defaults to all atoms}

\item{write_pdf}{Whether to generate a pdf report}

\item{write_csv}{Whether to generate csv tables}

\item{verbose}{Verbosity level: 0 = silent, 1 = normal, 2 = verbose}

\item{...}{Arguments passed on to \code{single_report}
\describe{
  \item{results_file}{The file from which to extract data}
  \item{results_files_dir}{The directory to save output data}
  \item{report_template}{The template for the report}
  \item{element_list}{The list of elements to check for}
  \item{write_pdf}{Binary: write a pdf file of the tables}
  \item{write_csv}{Binary: write the tables into a csv file (one each)}
  \item{verbose}{Level of information output: 0 - none, 1 - normal, 2 - high.}
}}
}
\description{
Compile Results & Generate a Report
}
