% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{element_count}
\alias{element_count}
\title{Element Counter}
\usage{
element_count(formula, element)
}
\arguments{
\item{formula}{A single molecular formula as a string (e.g. C12H26).}

\item{element}{The element of which you want the number of atoms.}
}
\value{
The number of atoms of the requested element in the formula
}
\description{
A simple wrapper for str_match regex to find the number of atoms of a provided \code{element} in a chemical \code{formula}. Returns the number of atoms, or 0 if not found.
}
\keyword{Internal}
