% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elements}
\alias{elements}
\title{Elements Symobol List}
\format{A character vector of length 119}
\usage{
elements
}
\description{
This list of elements includes all elements with IUPAC assigned symbols, plus Deuterium
}
\keyword{datasets}
