% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{extract_key_value}
\alias{extract_key_value}
\title{Extract Header Info}
\usage{
extract_key_value(info_frame, info_key)
}
\arguments{
\item{info_frame}{The data frame, with two columns: \code{key} and \code{value}}

\item{info_key}{The key from which to extract data, if it exists.}
}
\value{
value from key:value pair
}
\description{
Get values out of a data frame with key:value columns. Return the value if string, or stop() if not found
}
\keyword{Internal}
